+++ src/rai_core/rai/agents/langchain/__init__.py
@@ -14,7 +14,7 @@
 
 from .agent import BaseState, LangChainAgent, newMessageBehaviorType
 from .callback import HRICallbackHandler
-from .core.runnables import (
+from .core import (
     ReActAgentState,
     create_react_runnable,
     create_state_based_runnable,
