+++ src/rai_core/rai/agents/langchain/core/__init__.py
@@ -14,11 +14,11 @@
 
 from .conversational_agent import State as ConversationalAgentState
 from .conversational_agent import create_conversational_agent
-from .runnables import (
+from .react_agent import (
     ReActAgentState,
     create_react_runnable,
-    create_state_based_runnable,
 )
+from .state_based_agent import create_state_based_runnable
 from .tool_runner import ToolRunner
 
 __all__ = [
