+++ src/rai_core/rai/agents/langchain/core/state_based_agent.py
@@ -83,54 +83,6 @@ def llm_node(
     state["messages"].append(ai_msg)
 
 
-def create_react_runnable(
-    llm: Optional[BaseChatModel] = None,
-    tools: Optional[List[BaseTool]] = None,
-    system_prompt: Optional[str | SystemMultimodalMessage] = None,
-) -> Runnable[ReActAgentState, ReActAgentState]:
-    """Create a react agent that can process messages and optionally use tools.
-
-    Parameters
-    ----------
-    llm : Optional[BaseChatModel], default=None
-        Language model to use. If None, will use complex_model from config
-    tools : Optional[List[BaseTool]], default=None
-        List of tools the agent can use
-
-    Returns
-    -------
-    Runnable[ReActAgentState, ReActAgentState]
-        A runnable that processes messages and optionally uses tools
-
-    Raises
-    ------
-    ValueError
-        If tools are provided but invalid
-    """
-    if llm is None:
-        llm = get_llm_model("complex_model", streaming=True)
-
-    graph = StateGraph(ReActAgentState)
-    graph.add_edge(START, "llm")
-
-    if tools:
-        tool_runner = ToolRunner(tools)
-        graph.add_node("tools", tool_runner)
-        graph.add_conditional_edges(
-            "llm",
-            tools_condition,
-        )
-        graph.add_edge("tools", "llm")
-        # Bind tools to LLM
-        bound_llm = cast(BaseChatModel, llm.bind_tools(tools))
-        graph.add_node("llm", partial(llm_node, bound_llm, system_prompt))
-    else:
-        graph.add_node("llm", partial(llm_node, llm, system_prompt))
-
-    # Compile the graph
-    return graph.compile()
-
-
 def retriever_wrapper(
     state_retriever: Callable[[], Dict[str, HumanMessage | HumanMultimodalMessage]],
     state: ReActAgentState,
